requirements.txt
fastapi
uvicorn
firebase-admin
psycopg2
python-dotenv
passlib
requests
